Initial commit.
Used the "ship" schematic and adjusted that to match the hardware of the gunship. Used the ship.ino code and merged the HelicopterEffects code in there. The state machines previously used for DC motor revving where removed because the entire thing is now timer event based anyway (so use timer values instead of states)